+++ crates/nostr-relay-builder/src/local/inner.rs
@@ -18,7 +18,8 @@ use tokio::sync::{broadcast, Semaphore};
 use super::session::{Nip42Session, RateLimiterResponse, Session, Tokens};
 use super::util;
 use crate::builder::{
-    RateLimit, RelayBuilder, RelayBuilderMode, RelayBuilderNip42, RelayTestOptions,
+    PolicyResult, QueryPolicy, RateLimit, RelayBuilder, RelayBuilderMode, RelayBuilderNip42,
+    RelayTestOptions, WritePolicy,
 };
 use crate::error::Error;
 
@@ -39,6 +40,8 @@ pub(super) struct InnerLocalRelay {
     min_pow: Option<u8>, // TODO: use AtomicU8 to allow to change it?
     #[cfg(feature = "tor")]
     hidden_service: Option<String>,
+    write_policy: Vec<Arc<dyn WritePolicy>>,
+    query_policy: Vec<Arc<dyn QueryPolicy>>,
     nip42: Option<RelayBuilderNip42>,
     test: RelayTestOptions,
 }
@@ -102,6 +105,8 @@ impl InnerLocalRelay {
             min_pow: builder.min_pow,
             #[cfg(feature = "tor")]
             hidden_service,
+            write_policy: builder.write_plugins,
+            query_policy: builder.query_plugins,
             nip42: builder.nip42,
             test: builder.test,
         };
@@ -187,7 +192,7 @@ impl InnerLocalRelay {
                             match msg {
                                 Message::Text(json) => {
                                     tracing::trace!("Received {json}");
-                                    self.handle_client_msg(&mut session, &mut tx, ClientMessage::from_json(json)?)
+                                    self.handle_client_msg(&mut session, &mut tx, ClientMessage::from_json(json)?, &addr)
                                         .await?;
                                 }
                                 Message::Binary(..) => {
@@ -238,6 +243,7 @@ impl InnerLocalRelay {
         session: &mut Session,
         ws_tx: &mut WsTx,
         msg: ClientMessage,
+        addr: &SocketAddr,
     ) -> Result<()> {
         match msg {
             ClientMessage::Event(event) => {
@@ -305,6 +311,23 @@ impl InnerLocalRelay {
                                         "{}: you must auth",
                                         MachineReadablePrefix::AuthRequired
                                     ),
+                                },
+                            )
+                            .await;
+                    }
+                }
+
+                // check write policy
+                for policy in self.write_policy.iter() {
+                    let event_id = event.id;
+                    if let PolicyResult::Reject(m) = policy.admit_event(&event, addr).await {
+                        return self
+                            .send_msg(
+                                ws_tx,
+                                RelayMessage::Ok {
+                                    event_id,
+                                    status: false,
+                                    message: format!("{}: {}", MachineReadablePrefix::Blocked, m),
                                 },
                             )
                             .await;
@@ -502,6 +525,21 @@ impl InnerLocalRelay {
                                         "{}: you must auth",
                                         MachineReadablePrefix::AuthRequired
                                     ),
+                                },
+                            )
+                            .await;
+                    }
+                }
+
+                // check query policy plugins
+                for plugin in self.query_policy.iter() {
+                    if let PolicyResult::Reject(msg) = plugin.admit_query(&filters, addr).await {
+                        return self
+                            .send_msg(
+                                ws_tx,
+                                RelayMessage::Closed {
+                                    subscription_id,
+                                    message: format!("{}: {}", MachineReadablePrefix::Error, msg),
                                 },
                             )
                             .await;